[CHORE] 메일 전송 익스텐션으로 이동 (#91)
